backend/requirements.txt
sqlalchemy
Flask
argon2-cffi
pyjwt
black
